bugfix: queue log events before initialization (#171)
due to the order of effects (trigger in children before parents), its
simple to add a logevent in an effect underneath the statsig provider
which will cause problems - namely, the initialization will not have run
yet, so it will throw an error. This is undocumented, unexpected, and
logging should never through.

this introduces an event queueing system, so events that you try to log
before initialization are queued and logged on initialization, rather
than throwing an error.