[release] 1.38.1 - Bugfix - memoization in SynchronousProvider (#177)
We introduced a fix in 1.37.1 to memoize configs for the same update
time. We cannot memoize on the config update time, because it is both 1)
not set on server sdks when generating the initialize values, and 2) not
an exhaustive memoization key. Instead, we stringify the
initializeValues as the memoization key.

This update renames `initValuesTime` in the `StatsigContext` to force
anyone relying on this field to update their implementation due to these
problems.





>Included In This Release
>- ed06b9e0bedcfd4353a799430410a8d2e5bff77d tore-statsig
>   - bugfix: memoization in SynchronousProvider (#175)

---------

Co-authored-by: statsig-kong[bot] <statsig-kong[bot]@users.noreply.github.com>